+++ lib/src/apngframe.cpp
@@ -158,7 +158,7 @@ namespace apngasm {
           (void)png_set_interlace_handling(png_ptr);
           png_read_update_info(png_ptr, info_ptr);
           _colorType = png_get_color_type(png_ptr, info_ptr);
-          rowbytes = png_get_rowbytes(png_ptr, info_ptr);
+          rowbytes = (int)png_get_rowbytes(png_ptr, info_ptr);
           memset(_palette, 255, sizeof(_palette));
           memset(_transparency, 255, sizeof(_transparency));
